Revert "Updated Source Engine!"
This reverts commit f76a0e870f26452b662a204c5119ec0caa184fe4.

+++ src/game/client/imessagechars.h
@@ -21,7 +21,7 @@
 namespace vgui
 {
 class Panel;
-typedef uint32 HFont;
+typedef unsigned long HFont;
 }
 
 abstract_class IMessageChars
